+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.2.0]
+### Added
+- Search function for filtering resource lists
+  - Ability to filter by any string
+  - Ability to filter by "starts with" using ^
+
 ## [1.1.1] - 2019-06-22
 ### Added
 - Coverage via Codecov
@@ -31,7 +37,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Resource endpoint for resource lists
 - Unit tests for all endpoints
 
-[Unreleased]: https://github.com/mtslzr/pokeapi-go/compare/v1.1.1...HEAD
+[Unreleased]: https://github.com/mtslzr/pokeapi-go/compare/v1.2.0...HEAD
+[1.2.0]: https://github.com/mtslzr/pokeapi-go/compare/v1.1.1...v1.2.0
 [1.1.1]: https://github.com/mtslzr/pokeapi-go/compare/v1.1.0...v1.1.1
 [1.1.0]: https://github.com/mtslzr/pokeapi-go/compare/v1.0.0...v1.1.0
 [1.0.0]: https://github.com/mtslzr/pokeapi-go/releases/tag/v1.0.0